requirements.txt
boto3==1.26.152
channels-redis==4.1.0
click==8.1.3
daphne==4.0.0
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
fcm-django==2.0.0
grpcio-status==1.54.2
mysqlclient==2.1.1
pillow==9.5.0
pip-chill==1.0.3
pyopenssl==23.2.0
service-identity==23.1.0
slack-sdk==3.21.3
wsproto==1.2.0
